feat(translator): Phase 4.10 — compile_error feedback loop with naming retry
Re-prompt each compile_error from Phase 4.9-bis with the cc1 error
message plus a strict HELPER NAMING rule derived from xrefs_out
PascalCase. Pre-filter data-table mis-classifications (static const
without _c(Snes *snes) signature) so they are skipped rather than
wasted on retry.

Results on battle/:

  qwen3-coder:480b (62 compile_err -> 8 data_table_skip + 54 retried)
    pass            76 -> 88   (+12, +16%)
    compile_error   62 -> 39   (-23, -37%)
    custom_spike    11 -> 14   (+3)
    unknown         12 -> 22   (+10  — compile OK, runtime exit != 0)
    timeout          2 ->  0   (-2)
    pass rate     46.6% -> 54.0%

  gemma4:31b (34 compile_err -> 3 data_table_skip + 31 retried)
    pass            58 -> 64   (+6, +10%)
    compile_error   34 -> 21   (-13, -38%)
    custom_spike    55 -> 57   (+2)
    unknown         12 -> 19   (+7)
    timeout          2 ->  0   (-2)
    no_contract      3 ->  3   (0)
    pass rate     35.4% -> 39.0%

gemma4 now sits below the <15% compile_err threshold flagged in the
session-restart-guide as "industrial-grade". qwen3 still at 24% — a
second retry pass would have diminishing returns since the remaining
cases are structural rather than mechanical.

New: translator/retry_compile_errors.py
  - Reuses batch_translate helpers (hydrate, build_user_prompt,
    load_prompts) — no logic duplication.
  - FEEDBACK_SUFFIX = task prompt + previous C + cc1 error + strict
    HELPER NAMING rule (no snake_case, must keep exact PascalCase,
    canonical helpers list from xrefs_out).
  - Pre-retry filter _is_data_table() catches jump tables and other
    pure-data routines wrongly classified TRANSLATE upstream — saves
    11 wasted retries (8 qwen3 + 3 gemma4) and is honest about not
    being able to fix them.

Cost: \$0 (Ollama Pro absorbed), ~5 min retry + ~10 min validation.

Next: Phase 5 — scaffold external/ff4/ in retro-go-sd and validate
the STM32 cross-compile against the 88+ PASS translations.

+++ port/_runs/gemma4_31b/battle/GetTimerPtr.c
@@ -1,31 +1,27 @@
 // Entry mode: A 8-bit (mf=1), X 16-bit (xf=0), DB=$7E, DP=0
 // Logic:
-//   The routine calculates a pointer/value by adding a 16-bit constant from 
-//   WRAM [$3530-$3531] to the current value in the accumulator (A), 
-//   carrying over the overflow to the high byte.
-//   The result is stored in WRAM [$3598-$3599].
+//   Adds a 16-bit value from WRAM [$3530-$3531] to the 8-bit accumulator.
+//   The result is stored as a 16-bit word in WRAM [$3598-$3599].
 static void GetTimerPtr_c(Snes *snes) {
     uint8_t *ram = snes->ram;
-    uint16_t a = snes->cpu->a; // Mode A 8-bit: use low byte
+    uint8_t a = (uint8_t)snes->cpu->a;
 
     // clc / adc $3530
     uint16_t low_res = (uint16_t)a + ram[0x3530];
     ram[0x3598] = (uint8_t)low_res; // sta $3598
 
     // lda $3531 / adc #0 / sta $3599
-    // The carry from the previous addition is applied to the high byte
-    uint16_t carry = low_res >> 8;
-    uint16_t high_res = (uint16_t)ram[0x3531] + carry;
-    ram[0x3599] = (uint8_t)high_res; // sta $3599
+    uint8_t carry = (uint8_t)(low_res >> 8); // Pitfall 7: capture carry from 8-bit add
+    uint8_t high_res = (uint8_t)(ram[0x3531] + carry);
+    ram[0x3599] = high_res; // sta $3599
 }
 
-// PITFALLS: 7 (8-bit arithmetic truncation: result must be cast to uint8_t 
-// to match 65816 behavior and carry generation)
+// PITFALLS: 7 (8-bit arithmetic truncation used to derive carry for the high byte)
 // HELPERS: none
 // CONTRACT:
 //   inputs_reg:  a=8, x=none, y=none
 //   inputs_ram:  0x3530=1, 0x3531=1
-//   output_ram: 0x3598=2
+//   output_ram:  0x3598=2
 //   entry_mode:  mf=true, xf=false, dp=0x0, db=0x7E
 //   entry_flags: z=auto, n=auto
 REVERSED_FUNCTION: battle::GetTimerPtr ($85:69)
